+++ compiler/src/dotty/tools/dotc/decompiler/DecompilationPrinter.scala
@@ -9,8 +9,9 @@ import scala.io.Codec
 import dotty.tools.dotc.core.Contexts._
 import dotty.tools.dotc.core.Phases.Phase
 import dotty.tools.dotc.core.tasty.TastyPrinter
-import dotty.tools.dotc.quoted.QuoteContextImpl
 import dotty.tools.io.File
+
+import scala.quoted.internal.impl.QuoteContextImpl
 
 /** Phase that prints the trees in all loaded compilation units.
  *
