+++ compiler/src/scala/quoted/internal/impl/ScopeException.scala
@@ -1,3 +1,3 @@
-package dotty.tools.dotc.quoted
+package scala.quoted.internal.impl
 
 class ScopeException(msg: String) extends Exception(msg)
