+++ compiler/src/scala/quoted/internal/impl/printers/Extractors.scala
@@ -1,4 +1,5 @@
-package dotty.tools.dotc.quoted.printers
+package scala.quoted
+package internal.impl.printers
 
 import scala.quoted._
