+++ library/src/scala/quoted/internal/Expr.scala
@@ -3,9 +3,6 @@ package internal
 
 import scala.annotation.{Annotation, compileTimeOnly}
 
-/** Implementation of scala.quoted.Expr that sould only be extended by the implementation of `QuoteContext` */
-abstract class Expr[+T] extends scala.quoted.Expr[T]
-
 @compileTimeOnly("Illegal reference to `scala.quoted.internal.Expr`")
 object Expr:
